[chore] Rename newsletterTxtFormat.ts to newsletterFormat.ts
