+++ curator/src/mail_agent/newsletterFormat.ts
